+++ app/controllers/api/v1/fibonacci_controller.rb
@@ -2,7 +2,11 @@ class Api::V1::FibonacciController < Api::BaseController
 
   def create
     if permitted_fibonacci[:n]
-      @fibonacci = Struct.new(:id, :value, :result, :runtime).new(1, 10, 55, 67)
+      n = permitted_fibonacci[:n].to_i
+      start_time = Time.now
+      service = FibonacciAlgorithmService.call(n)
+      end_time = (Time.now - start_time) * 1000
+      @fibonacci = Struct.new(:id, :value, :result, :runtime).new(1, n, service.result, end_time)
       render json: @fibonacci.to_json, status: :created
     else
       render json: { error: "Please set the param 'n'" }, status: :bad_request