+++ spec/support/requests_helper.rb
@@ -1,7 +1,7 @@
 module Requests
   module JsonHelpers
     def json
-      JSON.parse(response.body)
+      @json ||= JSON.parse(response.body)
     end
   end
 end